# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MCP Aggregator",
    "description": "A centralized location for LLM applications to access external context, tools, and prompts",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/apps/{app}/tools/": {
      "get": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "List App Tools",
        "description": "List all tools available for a specific app\n\nArgs:\n    app (str): App name (e.g., 'github')\n    db (Session): Database session\n    \nReturns:\n    List[schemas.Tool]: List of tools for the app",
        "operationId": "list_app_tools_api_v1_apps__app__tools__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tool"
                  },
                  "title": "Response List App Tools Api V1 Apps  App  Tools  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/apps/{app}/tools/{tool}/": {
      "get": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Get Tool",
        "description": "Get a specific tool by app and tool name\n\nArgs:\n    app (str): App name (e.g., 'github')\n    tool (str): Tool name (e.g., 'github.list_repos')\n    db (Session): Database session\n    \nReturns:\n    schemas.Tool: Tool details\n    \nRaises:\n    HTTPException: If tool is not found",
        "operationId": "get_tool_api_v1_apps__app__tools__tool___get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "App"
            }
          },
          {
            "name": "tool",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guess-tools/": {
      "get": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Guess Tools",
        "operationId": "guess_tools_api_v1_guess_tools__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/execute/": {
      "post": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Execute Tool",
        "description": "Execute a tool with provided parameters\n\nArgs:\n    request (schemas.ExecuteToolRequest): Tool execution request\n    db (Session): Database session\n    current_user (models.User): Current authenticated user\n    \nReturns:\n    schemas.ExecuteToolResponse: Execution result\n    \nRaises:\n    HTTPException: If tool is not found or execution fails",
        "operationId": "execute_tool_api_v1_execute__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tools/{tool_id}/execute/": {
      "post": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Execute Specific Tool",
        "description": "Execute a tool by its ID\n\nArgs:\n    tool_id (int): Tool ID in the database\n    db (Session): Database session\n    current_user (models.User): Current authenticated user\n    \nReturns:\n    schemas.ExecuteToolResponse: Execution result",
        "operationId": "execute_specific_tool_api_v1_tools__tool_id__execute__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/execute/log/{execution_log_id}/rate/": {
      "post": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Rate Execution",
        "operationId": "rate_execution_api_v1_execute_log__execution_log_id__rate__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "execution_log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Execution Log Id"
            }
          },
          {
            "name": "rating",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rating"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/": {
      "get": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "List Tools",
        "operationId": "list_tools_api_v1_tools__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Tool"
                  },
                  "type": "array",
                  "title": "Response List Tools Api V1 Tools  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/mcp-url/": {
      "get": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Generate Mcp Url",
        "description": "Generate a unique URL for the user to connect their MCP client\n\nThis URL contains authentication information that allows the MCP client\nto access all tools available to the user without additional authentication.\n\nArgs:\n    db (Session): Database session\n    current_user (models.User): Current authenticated user\n    \nReturns:\n    schemas.MCPUrlResponse: MCP URL and associated metadata",
        "operationId": "generate_mcp_url_api_v1_mcp_url__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPUrlResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/tools/{tool_id}/": {
      "get": {
        "tags": [
          "tools",
          "tools"
        ],
        "summary": "Get Tool By Id",
        "operationId": "get_tool_by_id_api_v1_tools__tool_id___get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tool"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExecuteToolRequest": {
        "properties": {
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "parameters": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          }
        },
        "type": "object",
        "required": [
          "tool"
        ],
        "title": "ExecuteToolRequest"
      },
      "ExecuteToolResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "result": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ExecuteToolResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MCPUrlResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": "URL for connecting to the MCP server with your credentials"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "MCPUrlResponse"
      },
...
      "Tool": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "parameters": {
            "type": "object",
            "title": "Parameters"
          },
          "action_definition": {
            "type": "object",
            "title": "Action Definition"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "app_id": {
            "type": "integer",
            "title": "App Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "parameters",
          "action_definition",
          "id",
          "app_id"
        ],
        "title": "Tool"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}